+++ CHANGES.rst
@@ -4,6 +4,11 @@ Changes
 Unreleased
 ----------
 
+Version 0.6.0
+-------------
+
+Released 2021-05-04
+
 - Make decode accept multiple keys. `#61 <https://github.com/dajiaji/python-cwt/pull/61>`__
 - Add set_private_claim_names to ClaimsBuilder and CWT. `#60 <https://github.com/dajiaji/python-cwt/pull/60>`__
 - Add sample of CWT with user-defined claims to docs. `#60 <https://github.com/dajiaji/python-cwt/pull/60>`__